Move over SerialConsole and Novnc, remove harvester.notifications - These components were so 1-1 with harvester in the end i moved them over - Also removed the like-for-like generic.notification `harvester.notification` strings
