+++ pkg/harvester/components/SerialConsole/index.vue
