+++ pkg/harvester/components/novnc/NovncConsoleItem.vue
