+++ pkg/harvester/dialog/HarvesterAddHotplugModal.vue
@@ -83,7 +83,7 @@ export default {
 
           if (res._status === 200 || res._status === 204) {
             this.$store.dispatch('growl/success', {
-              title:   this.t('harvester.notification.title.succeed'),
+              title:   this.t('generic.notification.title.succeed'),
               message: this.t('harvester.modal.hotplug.success', { diskName: this.diskName, vm: this.actionResource.nameDisplay })
             }, { root: true });
